requirements.txt
asgiref==3.11.0
attrs==25.4.0
Django==4.1.13
django-scalar==0.1.4
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.29.0
hatchling==1.28.0
inflection==0.5.1
joblib==1.5.3
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
numpy==2.4.0
packaging==25.0
pathspec==0.12.1
pluggy==1.6.0
PyJWT==2.10.1
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.30.0
scikit-learn==1.8.0
scipy==1.16.3
sqlparse==0.5.5
threadpoolctl==3.6.0
trove-classifiers==2025.12.1.14
uritemplate==4.2.0
